+++ xchange-btcchina/src/main/java/com/xeiam/xchange/btcchina/dto/trade/request/BTCChinaGetIcebergOrderRequest.java
@@ -29,7 +29,7 @@ public class BTCChinaGetIcebergOrderRequest extends BTCChinaRequest {
 
   private static final String METHOD_NAME = "getIcebergOrder";
 
-  public BTCChinaGetIcebergOrderRequest(long id, String market) {
+  public BTCChinaGetIcebergOrderRequest(int id, String market) {
 
     this.method = METHOD_NAME;
     this.params = String.format("[%1$d,\"%2$s\"]", id, market == null ? BTCChinaExchange.DEFAULT_MARKET : market);
